simpilfy refresh code for routes on cf_app
The original code to read routes into the new "routes" block was overcomplicated
and tended to result in unintended failures during Terraform's refresh step.
The new code simply reads the current reality into state and relies on the
update step to sort it out.  Routes that are mapped to the app but that do not
appear in the state are still ignored to avoid conflicts with routes that are
mapped to the app manually or via the cf_route resource.

Also removed the routes.exclusive flag as per developer discussions.