Allow diff customization for changed `docker_image` and `path`
Fixes cloudfoundry-community/terraform-provider-cloudfoundry#311

When `docker_image` or `path` parameters of a `cloudfoundry_app` are changed from a non-empty value, customization rules should not be ignored.
For example, for a `docker_image` being updated in the context of a blue/green deployment, it is required for the `id_bg` property to be flagged as being re-computed to reflect the change of GUID which will be caused by the deployment.